Keep count of individual trackers for each (host,hash) pair. Closes #4267.
If the tracker list for a torrent includes several entries pointing to
the same host, removing a single entry should not immediately cause the
(host,torrent) item to disappear from the tracker filters list.

By keeping count of all actual tracker URLs for each (host,hash) pair,
we can determine whether or not a removed entry was the last one
remaining.